resnet_server/requirements.txt
flask
torch
torchvision
Pillow
requests
aiohttp
prometheus_client
redis